fix: use next.config.esm to make package work
