+++ website/package.json
@@ -20,14 +20,14 @@
     "autoprefixer": "10.4.14",
     "eslint": "8.40.0",
     "eslint-config-next": "13.4.3",
-    "next": "13.4.3",
+    "next": "^13.5.6",
     "postcss": "8.4.23",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "rehype-pretty-code": "^0.10.0",
+    "rehype-pretty-code": "^0.10.2",
     "shiki": "^0.14.2",
     "tailwindcss": "3.3.2",
-    "typescript": "5.0.4"
+    "typescript": "^5.2.2"
   },
   "browserslist": [
     "> 0.5%",
