+++ website/tsconfig.json
@@ -24,5 +24,6 @@
     }
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
+  "files": ["next.config.mjs"],
   "exclude": ["node_modules"]
 }
